fix : temp

+++ movie-community_spring-server/src/main/java/com/codeworrisors/Movie_Community_Web/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package com.codeworrisors.Movie_Community_Web.repository;
 import com.codeworrisors.Movie_Community_Web.model.Review;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ReviewRepository extends PagingAndSortingRepository<Review, Integer> {
+public interface ReviewRepository extends PagingAndSortingRepository<Review, Long> {
 }
